+++ client/Packages/com.beamable/CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Content fields with attribute `MustReferenceContent` will have display content picker inspector in Editor
+- Added a QA Smoke Kit sample (internal smoke-test harness).
 
 ### Fixed